feat: game transfer controllers
Inverts control flow for unit and resource transfers and assumes the
game is the source of truth for what can be transferred for these
subsystems. This allows the engine to step completely out of the way
and be very specific about what it needs from the game to do that.